backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.32
alembic==1.13.2
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.2.0
httpx==0.27.0
pydantic==2.8.2
pydantic-settings==2.4.0
PyMySQL==1.1.1
PyYAML==6.0.1
pytest==8.3.2
authlib==1.3.1
factory-boy==3.3.0
Faker==30.6.0
openpyxl==3.1.5
boto3>=1.35.40
google-genai==0.4.0
